requirements.txt
Flask
Werkzeug
SQLAlchemy
Flask-SQLAlchemy
Flask-WTF
WTForms
cryptography
web3
ipfshttpclient
tensorflow
keras
opencv-python
Pillow
numpy
scikit-learn
matplotlib
python-dotenv
